Move most of the core mods to core::tactical_map mod
