+++ src/core/tactical_map/ability.rs
@@ -1,5 +1,5 @@
 use core::map::Distance;
-use core::{Attacks, Strength};
+use core::tactical_map::{Attacks, Strength};
 
 /// Active ability.
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, Serialize, Deserialize)]
