+++ src/core/tactical_map/apply.rs
@@ -1,9 +1,9 @@
-use core::ability::{self, Ability};
-use core::component::{self, Component};
-use core::effect::{self, Effect, LastingEffect};
-use core::event::{self, ActiveEvent, Event};
-use core::state;
-use core::{Attacks, Jokers, Moves, ObjId, PlayerId, State};
+use core::tactical_map::ability::{self, Ability};
+use core::tactical_map::component::{self, Component};
+use core::tactical_map::effect::{self, Effect, LastingEffect};
+use core::tactical_map::event::{self, ActiveEvent, Event};
+use core::tactical_map::state;
+use core::tactical_map::{Attacks, Jokers, Moves, ObjId, PlayerId, State};
 
 pub fn apply(state: &mut State, event: &Event) {
     debug!("event::apply: {:?}", event);
