+++ src/core/tactical_map/effect.rs
@@ -1,5 +1,5 @@
-use core::component::Component;
-use core::{Phase, PosHex, Strength};
+use core::tactical_map::component::Component;
+use core::tactical_map::{Phase, PosHex, Strength};
 
 #[derive(Clone, Debug, Copy, PartialEq, Serialize, Deserialize)]
 pub enum Duration {
