+++ src/core/tactical_map/movement.rs
@@ -1,10 +1,10 @@
 use std::collections::VecDeque;
 use std::slice::Windows;
 
-use core::ability::PassiveAbility;
 use core::map::{dirs, Dir, Distance, HexMap, PosHex};
-use core::state;
-use core::{ObjId, State, TileType};
+use core::tactical_map::ability::PassiveAbility;
+use core::tactical_map::state;
+use core::tactical_map::{ObjId, State, TileType};
 
 #[derive(Serialize, Deserialize, Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord)]
 pub struct MovePoints(pub i32);
@@ -246,8 +246,8 @@ impl Pathfinder {
 
 #[cfg(test)]
 mod tests {
-    use core::movement::{Path, Step};
-    use core::PosHex;
+    use core::tactical_map::movement::{Path, Step};
+    use core::tactical_map::PosHex;
 
     const NODE_0: PosHex = PosHex { q: 0, r: 1 };
     const NODE_1: PosHex = PosHex { q: 1, r: 0 };
