+++ src/core/tactical_map/utils.rs
