+++ src/screen/battle/visualize.rs
@@ -6,12 +6,12 @@ use ggez::Context;
 use scene::action;
 use scene::{Action, Boxed, Sprite};
 
-use core::ability::Ability;
-use core::effect::{self, Effect, LastingEffect, TimedEffect};
-use core::event::{self, ActiveEvent, Event};
-use core::execute::ApplyPhase;
 use core::map::PosHex;
-use core::{ObjId, PlayerId, State};
+use core::tactical_map::ability::Ability;
+use core::tactical_map::effect::{self, Effect, LastingEffect, TimedEffect};
+use core::tactical_map::event::{self, ActiveEvent, Event};
+use core::tactical_map::execute::ApplyPhase;
+use core::tactical_map::{ObjId, PlayerId, State};
 use geom;
 use screen::battle::view::BattleView;
 use utils::time_s;
